+++ ufo/experience/config.yaml
@@ -1,0 +1,9 @@
+version: 0.1
+
+VISUAL_SYSTEM_PROMPT_TEMPLTE_FILE: ""
+VISUAL_SYSTEM_PROMPT_TEMPLTE_FILE: ""
+
+VISUAL_SYSTEM_PROMPT_TEMPLTE_FILE: ""
+VISUAL_SYSTEM_PROMPT_TEMPLTE_FILE: ""
+
+SAVE_PATH: ""
